refactor: split node logic
