+++ src/datafold_node/mod.rs
@@ -16,7 +16,9 @@ pub mod http_server;
 pub mod log_routes;
 pub mod network_routes;
 pub mod node;
+mod network;
 mod permissions;
+mod schema_manager;
 pub mod query_routes;
 pub mod schema_routes;
 pub mod security_routes;
